+++ app/utils/constants.py
@@ -9,4 +9,4 @@ KEYS_REQUESTED = (
     "time_limit",
     "memory_limit",
 )
-PROBLEMS_FOLDER = "../IO/problems/"
+PROBLEMS_FOLDER = "../mnt/problems/"
