+++ tests/conftest.py
@@ -2,8 +2,7 @@
 import copy
 from typing import Any
 import pytest
-from tests.resources import DATA_ENDPOINT
-
+from tests.resources import DATA_ENDPOINT, DATA_PROBLEM
 from app import create_app
 
 
@@ -32,3 +31,17 @@ def data_end_example():
         "memory_limit": 256,
         "language": "py3",
     }
+
+
+@pytest.fixture(name="app")
+def app_with_create_problem():
+    """Fixture providing app for testing problem endpoint"""
+    app = create_app()
+    app.config["TESTING"] = True
+    return app
+
+
+@pytest.fixture(name="data_problem")
+def data_example():
+    """Fixture providing example data for testing new_problem endpoint"""
+    return DATA_PROBLEM
