--- utils/constants.py
@@ -1,3 +1,0 @@
-"""This module contains constants for tests"""
-
-PROBLEMS_FOLDER = "../IO/problems/"
